Revert "moving common validation to base class (#773)"
This reverts commit 4142cfec7b436be3b999cf7c3fd5f78593c38308.
