+++ search-commons/src/main/java/no/unit/nva/search/common/enums/SortKey.java
@@ -15,8 +15,6 @@ import java.util.stream.Stream;
  */
 public interface SortKey {
 
-  String INVALID_SORT_KEY_NAME = "INVALID";
-
   static Predicate<SortKey> equalTo(String name) {
     return key -> name.matches(key.keyPattern());
   }
@@ -32,10 +30,6 @@ public interface SortKey {
   }
 
   String name();
-
-  default boolean isInvalid() {
-    return INVALID_SORT_KEY_NAME.equals(name());
-  }
 
   String keyPattern();
